rename LINCSDataset file.
